+++ ChangeLog.md
@@ -7,6 +7,9 @@ Next version
 =======================================================================================
 
 ### Added
+- Added `ranges` property to [`DateRange`](packages/moxb/src/date/DateRange.ts) and [`DateRangeImpl`](packages/moxb/src/date/DateRangeImpl.ts)
+  These are predetermined date range selectors allowing easy date-range selection. (i.e. current month)
+- Fixed the antd properties of [`DateRangePickerAnt`](packages/antd/src/DateRangePicker.tsx)
 
 ### Fixed